Задание 5. Выделение спецклассов помощников.
Выделила из ApplicationManager специализированные классы-помощники NavigationHelper, GroupHelper и SessionHelper, перенесла в них из ApplicationManager соответствующие вспомогательные методы.
