+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/appmanager/ApplicationManager.java
@@ -3,7 +3,6 @@ package ru.stqa.pft.addressbook.appmanager;
 import org.openqa.selenium.*;
 import org.openqa.selenium.firefox.FirefoxDriver;
 import ru.stqa.pft.addressbook.model.ContactData;
-import ru.stqa.pft.addressbook.model.GroupData;
 
 import java.util.concurrent.TimeUnit;
 
@@ -11,55 +10,21 @@ import static org.testng.Assert.assertTrue;
 import static org.testng.Assert.fail;
 
 public class ApplicationManager {
-  private WebDriver driver;
-  private String baseUrl;
+  protected WebDriver driver;
+  private NavigationHelper navigationHelper;
+  private GroupHelper groupHelper;
+  private SessionHelper sessionHelper;
   private boolean acceptNextAlert = true;
   private StringBuffer verificationErrors = new StringBuffer();
 
   public void init() {
     driver = new FirefoxDriver();
-    baseUrl = "https://www.google.com/";
     driver.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
     driver.get("http://localhost/addressbook/");
-    login("admin", "secret");
-  }
-
-  private void login(String username, String password) {
-    driver.findElement(By.name("user")).click();
-    driver.findElement(By.name("user")).clear();
-    driver.findElement(By.name("user")).sendKeys(username);
-    driver.findElement(By.name("pass")).click();
-    driver.findElement(By.name("pass")).clear();
-    driver.findElement(By.name("pass")).sendKeys(password);
-    driver.findElement(By.xpath("//input[@value='Login']")).click();
-  }
-
-  public void returnToGroupPage() {
-    driver.findElement(By.linkText("group page")).click();
-  }
-
-  public void submitGroupCreation() {
-    driver.findElement(By.name("submit")).click();
-  }
-
-  public void fillGroupForm(GroupData groupData) {
-    driver.findElement(By.name("group_name")).click();
-    driver.findElement(By.name("group_name")).clear();
-    driver.findElement(By.name("group_name")).sendKeys(groupData.name());
-    driver.findElement(By.name("group_header")).click();
-    driver.findElement(By.name("group_header")).clear();
-    driver.findElement(By.name("group_header")).sendKeys(groupData.header());
-    driver.findElement(By.name("group_footer")).click();
-    driver.findElement(By.name("group_footer")).clear();
-    driver.findElement(By.name("group_footer")).sendKeys(groupData.footer());
-  }
-
-  public void initGroupCreation() {
-    driver.findElement(By.name("new")).click();
-  }
-
-  public void gotoGroupPage() {
-    driver.findElement(By.linkText("groups")).click();
+    groupHelper = new GroupHelper(driver);
+    navigationHelper = new NavigationHelper(driver);
+    sessionHelper = new SessionHelper(driver);
+    sessionHelper.login("admin", "secret");
   }
 
   public void stop() {
@@ -103,14 +68,6 @@ public class ApplicationManager {
     }
   }
 
-  public void deleteSelectedGroups() {
-    driver.findElement(By.name("delete")).click();
-  }
-
-  public void selectGroup() {
-    driver.findElement(By.name("selected[]")).click();
-  }
-
   public void returnToHomepage() {
     driver.findElement(By.linkText("home")).click();
   }
@@ -148,5 +105,13 @@ public class ApplicationManager {
 
   public void selectContact() {
     driver.findElement(By.name("selected[]")).click();
+  }
+
+  public GroupHelper getGroupHelper() {
+    return groupHelper;
+  }
+
+  public NavigationHelper getNavigationHelper() {
+    return navigationHelper;
   }
 }
