+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/appmanager/SessionHelper.java
@@ -1,0 +1,20 @@
+package ru.stqa.pft.addressbook.appmanager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+
+public class SessionHelper {
+  private WebDriver driver;
+  public SessionHelper(WebDriver driver) {
+    this.driver = driver;
+  }
+  public void login(String username, String password) {
+    driver.findElement(By.name("user")).click();
+    driver.findElement(By.name("user")).clear();
+    driver.findElement(By.name("user")).sendKeys(username);
+    driver.findElement(By.name("pass")).click();
+    driver.findElement(By.name("pass")).clear();
+    driver.findElement(By.name("pass")).sendKeys(password);
+    driver.findElement(By.xpath("//input[@value='Login']")).click();
+  }
+}
